genAIsamu/requirements.txt
mistralai
streamlit
sqlite-utils
python-dotenv
requests